Add bf16 training, activation checkpointing, and SDPA multi-channel attention
Adds bf16 mixed-precision training/eval, optional upstream activation
checkpointing (--upstream_checkpoint_activations), an SDPA-backed attention
path (attn_impl) to avoid OOM in multi-channel cross-channel/cross-frame
layers, a learned channel attention-pooling layer (attnpool), and an
LFB+IPD2 feature type for the unix_enc_custom upstream. Also fixes dtype
handling in diar_ami loss/metrics and switches the default download cache
to a workspace path.

+++ s3prl/downstream/diar_ami/expert.py
@@ -200,7 +200,7 @@ class DownstreamExpert(nn.Module):
                 the loss to be optimized, should not be detached
         """
 
-        features = (pad_sequence(features, batch_first=True)).float()
+        features = pad_sequence(features, batch_first=True)
         labels = [torch.from_numpy(label).float() for label in labels]
         labels = pad_sequence(labels, batch_first=True).to(features.device)
         lengths = torch.tensor(lengths, dtype=torch.long)
